Set extra deps; Fix clients
