+++ src/concrete-async/concrete_async/models.py
@@ -1,10 +1,11 @@
 import json
 
-import concrete_core
 from kombu.utils.json import register_type
 from openai.types.chat import ChatCompletion
 from pydantic import BaseModel
 from pydantic.fields import Field
+
+import concrete
 
 
 class KombuMixin(BaseModel):
@@ -22,18 +23,18 @@ class KombuMixin(BaseModel):
         return super().__init_subclass__(**kwargs)
 
 
-class Operation(concrete_core.models.base.ConcreteModel, KombuMixin):
+class Operation(concrete.models.base.ConcreteModel, KombuMixin):
     client_name: str
     function_name: str
     arg_dict: dict[str, list | dict]
 
 
-class ConcreteChatCompletion(ChatCompletion, concrete_core.models.base.ConcreteModel, KombuMixin):
+class ConcreteChatCompletion(ChatCompletion, concrete.models.base.ConcreteModel, KombuMixin):
     message_format_name: str = Field(description="Response format to parse completion into")
 
     @property
-    def message(self) -> concrete_core.models.messages.Message:
-        message_format: type[concrete_core.models.messages.Message] = concrete_core.models.messages.Message.dereference(
+    def message(self) -> concrete.models.messages.Message:
+        message_format: type[concrete.models.messages.Message] = concrete.models.messages.Message.dereference(
             self.message_format_name
         )
         return message_format.model_validate(json.loads(self.choices[0].message.content or "{}"))
