+++ src/concrete-core/concrete/clients/__init__.py
@@ -1,0 +1,1 @@
+from .base import CLIClient, Client, LMClient, model_to_schema  # noqa
