+++ src/concrete-core/concrete/models/messages.py
@@ -107,8 +107,3 @@ class NodeSummary(Message):
 
 class NodeUUID(Message):
     node_uuid: str = Field(description="UUID of the node")
-
-
-class Param(Message):
-    name: str = Field(description="Name of the parameter")
-    value: str = Field(description="Value of the parameter")
