+++ src/concrete-core/concrete/webutils.py
@@ -1,9 +1,10 @@
 from typing import cast
 
-from concrete_core.utils import verify_jwt
 from fastapi import Request, status
 from fastapi.responses import JSONResponse
 from starlette.middleware.base import BaseHTTPMiddleware
+
+from concrete.utils import verify_jwt
 
 
 class AuthMiddleware(BaseHTTPMiddleware):
