+++ tests/test_tools.py
@@ -2,9 +2,10 @@ import unittest
 from unittest.mock import Mock, patch
 
 import pytest
-from concrete_core.models import messages
-from concrete_core.tools import HTTPTool, invoke_tool
 from requests.exceptions import HTTPError
+
+from concrete.models import messages
+from concrete.tools import HTTPTool, invoke_tool
 
 
 def test_http_tool_process_response_ok():
@@ -25,7 +26,7 @@ def test_http_tool_process_response_not_ok():
         HTTPTool._process_response(mock_response, '')
 
 
-@patch('concrete_core.tools.HTTPClient')
+@patch('concrete.tools.HTTPClient')
 def test_http_tool_request(mock_http_client):
     mock_response = Mock()
     mock_response.ok = True
@@ -57,7 +58,7 @@ def test_http_tool_request(mock_http_client):
 
 
 class TestInvokeTool(unittest.TestCase):
-    @patch('concrete_core.tools.TOOLS_REGISTRY', new_callable=dict)
+    @patch('concrete.tools.TOOLS_REGISTRY', new_callable=dict)
     def test_invoke_tool_success(self, mock_tools_registry):
         self.mock_tool = Mock()
         self.mock_tool.mock_function = Mock(return_value='example result')
@@ -81,7 +82,7 @@ class TestInvokeTool(unittest.TestCase):
             )
             invoke_tool(tool)
 
-    @patch('concrete_core.tools.TOOLS_REGISTRY', new_callable=dict)
+    @patch('concrete.tools.TOOLS_REGISTRY', new_callable=dict)
     def test_invoke_tool_attributeerror(self, mock_tools_registry):
         self.mock_tool = Mock(spec=[])
         mock_tools_registry['mock_tool'] = self.mock_tool
@@ -93,7 +94,7 @@ class TestInvokeTool(unittest.TestCase):
         with self.assertRaises(AttributeError):
             invoke_tool(tool)
 
-    @patch('concrete_core.tools.TOOLS_REGISTRY', new_callable=dict)
+    @patch('concrete.tools.TOOLS_REGISTRY', new_callable=dict)
     def test_invoke_tool_typeerror(self, mock_tools_registry):
         self.mock_tool = Mock()
         self.mock_tool.mock_function = Mock(return_value='example result')
